# src/Medley.Web/wwwroot/swagger/v1/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Medley API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/fragments": {
      "get": {
        "tags": [
          "FragmentsApi"
        ],
        "operationId": "FragmentsApi_GetAll",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "x-position": 1
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FragmentDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/fragments/{id}": {
      "get": {
        "tags": [
          "FragmentsApi"
        ],
        "operationId": "FragmentsApi_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FragmentDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/fragments/by-source/{sourceId}": {
      "get": {
        "tags": [
          "FragmentsApi"
        ],
        "operationId": "FragmentsApi_GetBySourceId",
        "parameters": [
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FragmentDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/fragments/search": {
      "get": {
        "tags": [
          "FragmentsApi"
        ],
        "operationId": "FragmentsApi_Search",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FragmentSearchResult"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/api/fragments/titles": {
      "post": {
        "tags": [
          "FragmentsApi"
        ],
        "operationId": "FragmentsApi_GetTitles",
        "requestBody": {
          "x-name": "ids",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FragmentTitleDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "FragmentDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "categoryIcon": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string"
          },
          "sourceId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "sourceName": {
            "type": "string",
            "nullable": true
          },
          "sourceType": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/SourceType"
              }
            ]
          },
          "sourceDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "confidence": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConfidenceLevel"
              }
            ]
          },
          "confidenceComment": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "SourceType": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Unknown",
          "Meeting"
        ],
        "enum": [
          "Unknown",
          "Meeting"
        ]
      },
      "ConfidenceLevel": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Unclear",
          "Low",
          "Medium",
          "High",
          "Certain"
        ],
        "enum": [
          "Unclear",
          "Low",
          "Medium",
          "High",
          "Certain"
        ]
      },
      "FragmentSearchResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "categoryIcon": {
            "type": "string",
            "nullable": true
          },
          "sourceId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "sourceName": {
            "type": "string",
            "nullable": true
          },
          "sourceType": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/SourceType"
              }
            ]
          },
          "sourceDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "confidence": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConfidenceLevel"
              }
            ]
          },
          "confidenceComment": {
            "type": "string",
            "nullable": true
          },
          "similarity": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "FragmentTitleDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
    }
  }
}